+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 *
 
+## [0.6.2] - 2021-05-03
+
+### Added
+
+* Support for more information in order notes
+
 ## [0.6.1] - 2021-07-05
 
 ### Added